remove previous dirfs tests
